moved out url builder into util, moved state to root
