+++ stylua.toml
@@ -1,5 +1,6 @@
 indent_type = "Spaces"
 indent_width = 2
-column_width = 120
+column_width = 140
+
 [sort_requires]
 enabled = true
